split handlers into individual modules
