+++ .travis.yml
@@ -7,9 +7,9 @@ cache:
     - $HOME/.cache/pip
 install:
 - pip install -r requirements-dev.txt
-- pip install .
+- pip install -r requirements.txt
 script:
-- py.test -v tests/
+- PYTHONPATH=`pwd` py.test -v tests/
 - flake8 --count grouper/
 notifications:
   slack:
