+++ grouper/fe/handlers/group_enable.py
@@ -1,0 +1,21 @@
+from grouper.fe.util import GrouperHandler
+from grouper.models import AuditLog, Group
+
+
+class GroupEnable(GrouperHandler):
+    def post(self, group_id=None, name=None):
+        group = Group.get(self.session, group_id, name)
+        if not group:
+            return self.notfound()
+
+        members = group.my_members()
+        if not self.current_user.my_role(members) in ("owner", "np-owner"):
+            return self.forbidden()
+
+        group.enable()
+        self.session.commit()
+
+        AuditLog.log(self.session, self.current_user.id, 'enable_group',
+                     'Enabled group.', on_group_id=group.id)
+
+        return self.redirect("/groups/{}?refresh=yes".format(group.name))
