+++ grouper/fe/handlers/group_requests.py
@@ -1,0 +1,29 @@
+from grouper.fe.util import GrouperHandler
+from grouper.models import Group, REQUEST_STATUS_CHOICES, Request
+
+
+class GroupRequests(GrouperHandler):
+    def get(self, group_id=None, name=None):
+        group = Group.get(self.session, group_id, name)
+        if not group:
+            return self.notfound()
+
+        status = self.get_argument("status", None)
+        offset = int(self.get_argument("offset", 0))
+        limit = int(self.get_argument("limit", 100))
+        if limit > 9000:
+            limit = 9000
+
+        requests = group.my_requests(status).order_by(
+            Request.requested_at.desc()
+        )
+        members = group.my_members()
+
+        total = requests.count()
+        requests = requests.offset(offset).limit(limit)
+
+        self.render(
+            "group-requests.html", group=group, requests=requests,
+            members=members, status=status, statuses=REQUEST_STATUS_CHOICES,
+            offset=offset, limit=limit, total=total
+        )
