+++ grouper/fe/handlers/not_found.py
@@ -1,0 +1,6 @@
+from grouper.fe.util import GrouperHandler
+
+
+class NotFound(GrouperHandler):
+    def get(self):
+        return self.notfound()
